refactor : Menu > Command 수정 & 키워드 추가, 메소드 네이밍 수정
